This commit was manufactured by cvs2svn to create branch 'STLPORT_5_0'.
git-svn-id: https://stlport.svn.sourceforge.net/svnroot/stlport/branches/STLPORT_5_0/STLport@743 01b27279-081b-0410-8cf9-d9d9e0fb0389

+++ test/regression/copy4.cpp
@@ -18,7 +18,7 @@ int copy4_test(int, char**)
   cout<<"Results of copy4_test:"<<endl;
   typedef vector<int> IVec;
   vector<int> v1(10);
-  for(int loc = 0; loc < v1.size(); loc++)
+  for(size_t loc = 0; loc < v1.size(); loc++)
     v1[loc] = loc;
   vector<int> v2;
   insert_iterator<IVec> i(v2, v2.begin());
